Refactor: Move vectorization managers to sr_detection directory
