+++ src/training/steps/market_analysis/sr_detection/unified_vectorization_manager.py
