+++ src/training/steps/market_analysis/sr_detection/vectorbt_rolling_optimizer.py
